mmc: core: Use maximum timeout values in case TAAC field is 0
When plugging the broken card at MMC socket, we get the following oops:

,----[ Kernel Oops ]
| ...
| mmc1: new high speed SD card at address b368
| mmcblk1: mmc1:b368 SMI 1.87 GiB
| mmcblk1:
| kernel BUG at drivers/mmc/host/mxs-mmc.c:465!
| Unable to handle kernel NULL pointer dereference at virtual address 00000000
| pgd = c0004000
| [00000000] *pgd=00000000
| Internal error: Oops: 817 [#1] PREEMPT
| last sysfs file: /sys/kernel/uevent_seqnum
| ...
`----

The error is due to an invalid field in CSD register of this 2GB SD
card. The CSD version of this card is 1.0 but the TAAC field is an
invalid value 0. Since the kernel is making use of this TAAC field to
calculate the SD card timeout, an invalid value 0 can lead to a oops
error at mxs_ns_to_ssp_ticks().

This error is only found on this <= 2GB SD card. > 2GB SDHC cards will
not see this error because these cards are CSD version 2.0. For these
cards, kernel always use a fix timeout and calculation that causes the
oops error is skipped. Of course, <= 2GB SD card that fully complying
to SD card physical layer specification will not see this error
neither.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>